+++ src/graph3d/graph3d.js
@@ -2431,7 +2431,7 @@ G3DpreventDefault = function (event) {
  * @param {Number} y
  * @param {Number} z
  */
-Point3d = function(x, y, z) {
+function Point3d(x, y, z) {
   this.x = x !== undefined ? x : 0;
   this.y = y !== undefined ? y : 0;
   this.z = z !== undefined ? z : 0;